test: cleanup test files, keep only GPU test script
