--- minimax_m1_config.json
@@ -1,106 +1,0 @@
-{
-  "model_type": "minimax_m1",
-  "architectures": [
-    "MiniMaxM1ForCausalLM"
-  ],
-  "hidden_size": 6144,
-  "intermediate_size": 9216,
-  "num_hidden_layers": 80,
-  "num_attention_heads": 64,
-  "num_key_value_heads": 8,
-  "head_dim": 128,
-  "vocab_size": 200064,
-  "rope_theta": 10000000,
-  "max_position_embeddings": 10240000,
-  "rms_norm_eps": 1e-05,
-  "hidden_act": "silu",
-  "num_local_experts": 32,
-  "num_experts_per_tok": 2,
-  "attn_type_list": [
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    1
-  ],
-  "pad_token_id": 0,
-  "bos_token_id": 1,
-  "eos_token_id": 2,
-  "tie_word_embeddings": false,
-  "postnorm": true
-}
