--- minimax_m1_small_config.json
@@ -1,23 +1,0 @@
-{
-  "model_type": "minimax_m1",
-  "architectures": ["MiniMaxM1ForCausalLM"],
-  "hidden_size": 512,
-  "intermediate_size": 1024,
-  "num_hidden_layers": 4,
-  "num_attention_heads": 4,
-  "num_key_value_heads": 2,
-  "head_dim": 128,
-  "vocab_size": 1000,
-  "num_local_experts": 4,
-  "num_experts_per_tok": 2,
-  "attn_type_list": [0, 0, 0, 0, 0, 0, 0, 1],
-  "max_position_embeddings": 512,
-  "rms_norm_eps": 1e-6,
-  "pad_token_id": 0,
-  "bos_token_id": 1,
-  "eos_token_id": 2,
-  "tie_word_embeddings": false,
-  "use_fused_rope": false,
-  "use_dynamic_ntk": true,
-  "use_logn_attn": false
-}
