--- test_discover_minimax.py
@@ -1,198 +1,0 @@
-# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
-
-"""
-MiniMax-M1 模块发现测试
-
-这个测试验证 MiniMax-M1 模块能否被 FastDeploy 发现和注册
-"""
-
-import sys
-import os
-import importlib.util
-
-# Windows 临时目录修复
-if sys.platform == 'win32':
-    if 'TEMP' not in os.environ:
-        os.environ['TEMP'] = 'C:\\Temp'
-    os.makedirs(os.environ['TEMP'], exist_ok=True)
-    
-    # 创建一个临时目录模拟 /tmp
-    tmp_dir = os.environ['TEMP']
-    import tempfile
-    tempfile.tempdir = tmp_dir
-
-print("=" * 60)
-print("MiniMax-M1 模块发现测试")
-print("=" * 60)
-
-# 测试1: 检查模型文件存在
-print("\n[测试 1] 检查模型文件")
-model_file = "fastdeploy/model_executor/models/minimax_m1.py"
-if os.path.exists(model_file):
-    print(f"  [OK] {model_file} 存在")
-    size = os.path.getsize(model_file)
-    print(f"       大小: {size} bytes")
-else:
-    print(f"  [FAIL] {model_file} 不存在")
-    sys.exit(1)
-
-# 测试2: 检查语法
-print("\n[测试 2] 检查 Python 语法")
-try:
-    with open(model_file, "r", encoding="utf-8") as f:
-        code = f.read()
-    
-    compile(code, model_file, 'exec')
-    print("  [OK] 语法正确")
-except SyntaxError as e:
-    print(f"  [FAIL] 语法错误: {e}")
-    sys.exit(1)
-
-# 测试3: 检查类定义
-print("\n[测试 3] 检查类定义")
-required_classes = [
-    "MiniMaxM1ForCausalLM",
-    "MiniMaxM1Model",
-    "MiniMaxM1DecoderLayer",
-    "MiniMaxM1LightningAttention",
-    "MiniMaxM1StandardAttention",
-    "MiniMaxM1SparseMoEBlock",
-    "MiniMaxM1MLP",
-]
-
-for cls in required_classes:
-    if f"class {cls}" in code:
-        print(f"  [OK] {cls}")
-    else:
-        print(f"  [FAIL] {cls} 未找到")
-
-# 测试4: 检查 ModelRegistry 注册
-print("\n[测试 4] 检查 ModelRegistry 注册")
-if "@ModelRegistry.register_model_class" in code:
-    print("  [OK] @ModelRegistry.register_model_class 装饰器存在")
-else:
-    print("  [WARN] 未找到装饰器（可能手动注册）")
-
-# 测试5: 尝试动态加载模块
-print("\n[测试 5] 动态加载模块")
-try:
-    # 创建模块规范
-    spec = importlib.util.spec_from_file_location("minimax_m1", model_file)
-    
-    # 创建模块
-    module = importlib.util.module_from_spec(spec)
-    
-    # 设置依赖模块的 mock
-    sys.modules['fastdeploy'] = type(sys)('fastdeploy')
-    sys.modules['fastdeploy.config'] = type(sys)('fastdeploy.config')
-    sys.modules['fastdeploy.model_executor'] = type(sys)('fastdeploy.model_executor')
-    sys.modules['fastdeploy.model_executor.layers'] = type(sys)('fastdeploy.model_executor.layers')
-    sys.modules['fastdeploy.model_executor.layers.embeddings'] = type(sys)('fastdeploy.model_executor.layers.embeddings')
-    sys.modules['fastdeploy.model_executor.layers.lm_head'] = type(sys)('fastdeploy.model_executor.layers.lm_head')
-    sys.modules['fastdeploy.model_executor.layers.normalization'] = type(sys)('fastdeploy.model_executor.layers.normalization')
-    sys.modules['fastdeploy.model_executor.layers.rotary_embedding'] = type(sys)('fastdeploy.model_executor.layers.rotary_embedding')
-    sys.modules['fastdeploy.model_executor.layers.activation'] = type(sys)('fastdeploy.model_executor.layers.activation')
-    sys.modules['fastdeploy.model_executor.layers.linear'] = type(sys)('fastdeploy.model_executor.layers.linear')
-    sys.modules['fastdeploy.model_executor.layers.moe'] = type(sys)('fastdeploy.model_executor.layers.moe')
-    sys.modules['fastdeploy.model_executor.layers.moe.moe'] = type(sys)('fastdeploy.model_executor.layers.moe.moe')
-    sys.modules['fastdeploy.model_executor.layers.attention'] = type(sys)('fastdeploy.model_executor.layers.attention')
-    sys.modules['fastdeploy.model_executor.layers.attention.attention'] = type(sys)('fastdeploy.model_executor.layers.attention.attention')
-    sys.modules['fastdeploy.model_executor.models'] = type(sys)('fastdeploy.model_executor.models')
-    sys.modules['fastdeploy.model_executor.models.model_base'] = type(sys)('fastdeploy.model_executor.models.model_base')
-    sys.modules['paddle'] = type(sys)('paddle')
-    sys.modules['paddle.nn'] = type(sys)('paddle.nn')
-    sys.modules['paddle.nn.functional'] = type(sys)('paddle.nn.functional')
-    sys.modules['paddleformers'] = type(sys)('paddleformers')
-    sys.modules['paddleformers.transformers'] = type(sys)('paddleformers.transformers')
-    sys.modules['paddleformers.utils'] = type(sys)('paddleformers.utils')
-    sys.modules['paddleformers.utils.log'] = type(sys)('paddleformers.utils.log')
-    
-    # 添加必要的 mock 类
-    class MockFDConfig:
-        def __init__(self):
-            self.model_config = type('ModelConfig', (), {})()
-            self.parallel_config = type('ParallelConfig', (), {})()
-            self.parallel_config.tensor_parallel_size = 1
-            self.parallel_config.expert_parallel_size = 1
-            
-    class MockModelCategory:
-        TEXT_GENERATION = 1
-        
-    class MockModelRegistry:
-        _registry = {}
-        
-        @classmethod
-        def register_model_class(cls, **kwargs):
-            def decorator(fn):
-                cls._registry[kwargs.get('architecture', fn.__name__)] = fn
-                return fn
-            return decorator
-        
-        @classmethod
-        def get_class(cls, name):
-            return cls._registry.get(name)
-    
-    # 设置 mock
-    sys.modules['fastdeploy'].config = type('config', (), {'FDConfig': MockFDConfig})()
-    sys.modules['fastdeploy.model_executor.models.model_base'].ModelRegistry = MockModelRegistry
-    sys.modules['fastdeploy.model_executor.models.model_base'].ModelCategory = MockModelCategory
-    sys.modules['paddle'].no_grad = lambda: None
-    sys.modules['paddle'].Tensor = type('Tensor', (), {})
-    
-    # 尝试加载模块
-    try:
-        spec.loader.exec_module(module)
-        print("  [OK] 模块加载成功")
-        
-        # 检查类
-        for cls_name in required_classes:
-            if hasattr(module, cls_name):
-                print(f"       - {cls_name}: OK")
-            else:
-                print(f"       - {cls_name}: NOT FOUND")
-                
-    except Exception as e:
-        print(f"  [WARN] 模块加载部分失败: {type(e).__name__}: {e}")
-        print("         这是预期的，因为缺少完整的 FastDeploy 环境")
-
-except Exception as e:
-    print(f"  [ERROR] {type(e).__name__}: {e}")
-
-# 测试6: 检查测试文件
-print("\n[测试 6] 检查测试文件")
-test_files = [
-    "tests/model_executor/test_minimax_m1.py",
-    "test_minimax_m1_local.py",
-]
-for tf in test_files:
-    if os.path.exists(tf):
-        print(f"  [OK] {tf}")
-    else:
-        print(f"  [FAIL] {tf} 不存在")
-
-# 测试7: 检查配置文件
-print("\n[测试 7] 检查配置文件")
-config_file = "minimax_m1_config.json"
-if os.path.exists(config_file):
-    import json
-    with open(config_file, "r") as f:
-        config = json.load(f)
-    print(f"  [OK] {config_file} 存在")
-    print(f"       model_type: {config.get('model_type')}")
-    print(f"       hidden_size: {config.get('hidden_size')}")
-    print(f"       num_hidden_layers: {config.get('num_hidden_layers')}")
-else:
-    print(f"  [FAIL] {config_file} 不存在")
-
-print("\n" + "=" * 60)
-print("测试完成!")
-print("=" * 60)
-print("\n总结:")
-print("  - MiniMax-M1 模型代码已完成")
-print("  - 测试文件已创建")
-print("  - 配置文件已生成")
-print("  - 模块可以被 Python 正确解析")
-print("\n注意: 完整加载模型需要:")
-print("  1. 实际的模型权重文件")
-print("  2. 完整的 PaddlePaddle + FastDeploy 环境")
-print("=" * 60)
