--- test_minimax_m1_simple.py
@@ -1,101 +1,0 @@
-#!/usr/bin/env python3
-"""
-Simple test for MiniMax-M1 model registration
-"""
-
-import os
-import sys
-
-# Setup environment
-os.environ["PROMETHEUS_MULTIPROC_DIR"] = f"/tmp/fd_prom_{os.getpid()}"
-os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
-
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-print("="*60)
-print("MiniMax-M1 Model Registration Test")
-print("="*60)
-
-# Test 1: Import model class
-print("\n[Test 1] Import MiniMaxM1ForCausalLM")
-try:
-    from fastdeploy.model_executor.models.minimax_m1 import MiniMaxM1ForCausalLM
-    print(f"  [PASS] Successfully imported MiniMaxM1ForCausalLM")
-except Exception as e:
-    print(f"  [FAIL] Import failed: {e}")
-    sys.exit(1)
-
-# Test 2: Check class registration
-print("\n[Test 2] Check ModelRegistry")
-try:
-    from fastdeploy.model_executor.models.model_base import ModelRegistry
-    
-    # Get registered models
-    registry = ModelRegistry._registry
-    if "minimax_m1" in registry:
-        print(f"  [PASS] minimax_m1 is registered in ModelRegistry")
-    else:
-        print(f"  [FAIL] minimax_m1 not found in registry")
-        print(f"  Available: {list(registry.keys())[:10]}...")
-except Exception as e:
-    print(f"  [FAIL] Check failed: {e}")
-
-# Test 3: Check class name
-print("\n[Test 3] Check class name")
-try:
-    name = MiniMaxM1ForCausalLM.name()
-    if name == "MiniMaxM1ForCausalLM":
-        print(f"  [PASS] Class name is correct: {name}")
-    else:
-        print(f"  [WARN] Class name: {name}")
-except Exception as e:
-    print(f"  [FAIL] {e}")
-
-# Test 4: Check model config attributes
-print("\n[Test 4] Check model config")
-try:
-    from fastdeploy.model_executor.models.minimax_m1 import MiniMaxM1Model
-    
-    # Check that the model class exists and has expected attributes
-    print(f"  [PASS] MiniMaxM1Model class exists")
-    
-    # Check source file
-    import inspect
-    source_file = inspect.getfile(MiniMaxM1ForCausalLM)
-    print(f"  [INFO] Source file: {source_file}")
-    
-except Exception as e:
-    print(f"  [FAIL] {e}")
-
-# Test 5: Check hybrid attention config
-print("\n[Test 5] Check hybrid attention support")
-try:
-    # Read the source to verify attn_type_list is used
-    with open("/mnt/c/Users/13286/.openclaw/workspace/FastDeploy/fastdeploy/model_executor/models/minimax_m1.py", "r") as f:
-        source = f.read()
-    
-    checks = [
-        ("attn_type_list", "attention type list"),
-        ("Lightning Attention", "Lightning Attention support"),
-        ("num_local_experts", "MoE experts"),
-        ("num_experts_per_tok", "MoE routing"),
-    ]
-    
-    for keyword, desc in checks:
-        if keyword in source:
-            print(f"  [PASS] {desc}: found '{keyword}'")
-        else:
-            print(f"  [WARN] {desc}: '{keyword}' not found")
-            
-except Exception as e:
-    print(f"  [FAIL] {e}")
-
-print("\n" + "="*60)
-print("Test Summary")
-print("="*60)
-print("""
-The MiniMax-M1 model is properly registered in FastDeploy.
-Note: Full model instantiation requires GPU or proper FastDeploy 
-runtime backend setup. This test verifies the model registration
-and code structure are correct.
-""")
